♻️ refactor(server): replace MastraMarkdownEmbeddingProvider with MastraMarkdownEmbeddingsGeneratorProvider
